% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_util.R
\name{convert2}
\alias{convert2}
\title{Convert Specific Values in a Data Frame}
\usage{
convert2(data, from = "", to = "")
}
\arguments{
\item{data}{A data frame where value replacement will be performed.}

\item{from}{Original value to be replaced.}

\item{to}{New value that will replace the original value.}
}
\value{
A modified data frame with specified value replacements.
}
\description{
Replaces all occurrences of a specified value with another value within a data frame.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c("", "y", "z"))
convert2(df, from = "", to = NA)

}
