% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_util.R
\name{my_round}
\alias{my_round}
\title{Format numeric values to specified decimal places}
\usage{
my_round(x, digit = 0)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{digit}{Number of decimal places to display, 0 by default.}
}
\value{
A character vector of formatted values.
}
\description{
Converts numeric values to character strings with fixed decimal formatting using \code{sprintf()}.
}
\examples{
my_round(3.1415, 2)
my_round(3, 2)
}
