% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysummary.R
\name{tidysummary}
\alias{tidysummary}
\title{Create a tidy summary of a data frame}
\usage{
tidysummary(x, size = "small", digits = 2)
}
\arguments{
\item{x}{a \code{\link[base:data.frame]{data.frame}} to generate a summary for.
Can also be an \link[base:atomic]{atomic} vector (numeric, factor, boolean etc.).}

\item{size}{how large the summary should be. Either \code{small}, \code{medium}
or \code{large} (or simply \code{s},  \code{m} or \code{l}).}

\item{digits}{the number of decimal places for rounding. Can be
an integer or \code{NULL} (no rounding).}
}
\value{
A data frame with one row per column.
}
\description{
Create a summary (small, medium or large) of a data frame.
}
\details{
The returned data frame has the following columns:

\tabular{ll}{
  \code{column} \tab Name of the column \cr
  \code{n} \tab Number of observations \cr
  \code{class} \tab Data type \cr
  \code{min} \tab Minimum value \cr
  \code{max} \tab Maximum value \cr
  \code{mean} \tab Mean value \cr
  \code{sd} \tab Standard deviation \cr
  \code{unique} \tab Number of unique values (includes \code{NA}) \cr
  \code{median} \tab Median value \cr
  \code{na} \tab Number of \code{NA}s (missing values) \cr
  \code{mode} \tab Most frequent value \cr
  \code{quant25} \tab 25 percent quantile \cr
  \code{quant75} \tab 75 percent quantile \cr
  \code{levels} \tab Number of factor levels (only relevant for factors) \cr
  \code{head} \tab Top (first) value of the data frame \cr
  \code{tail} \tab Bottom (last) value of the data frame \cr
  \code{se} \tab Standard error \cr
  \code{skewness} \tab Skewness \cr
  \code{kurtosis} \tab Kurtosis \cr
  \code{bytes} \tab Size in bytes of the memory usage of the column
}

Note:

\itemize{
  \item For character vectors, the text length will be used to calculate the
        statistics.
  \item Missing values will be automatically removed before any statistical
        summaries are calculated
  \item Statistical summaries are produced using base R functions with
        defaults. Kurtosis and skewness is produced using code adapted
        from the moments package
        (\url{https://cran.r-project.org/web/packages/moments/index.html}).
}
}
\examples{
# Summary for data frame
tidysummary(iris)
tidysummary(iris, "medium")
tidysummary(iris, digits=2)

# Summary for atomic vector
tidysummary(iris$Sepal.Length, "large")

# Convert summary to Markdown
output <- tidysummary(iris)
to_markdown(output)

}
\seealso{
\code{\link[tidysummary:to_markdown]{to_markdown()}},
\code{\link[tidysummary:to_latex]{to_latex()}},
\code{\link[tidysummary:to_html]{to_html()}},
\code{\link[tidysummary:to_json]{to_json()}},
\code{\link[tidysummary:to_csv]{to_csv()}},
\code{\link[tidysummary:to_excel]{to_excel()}}
}
