% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_csv}
\alias{to_csv}
\title{Create a CSV file from a tidysummary}
\usage{
to_csv(df, file, overwrite = FALSE)
}
\arguments{
\item{df}{a data frame.}

\item{file}{filename to create on disk.}

\item{overwrite}{whether to overwrite the file if it already exists.}
}
\description{
Takes a tidysummary object and saves it into a CSV file with UTF-8 encoding.
}
\details{
Uses the \code{\link[utils:write.csv]{write.csv}} function to save the file.
}
\examples{
\dontrun{
data <- tidysummary(iris)

to_csv(data, file="filename.csv")
}
}
\seealso{
\code{\link[tidysummary:tidysummary]{tidysummary()}}
}
