% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_json}
\alias{to_json}
\title{Create a table in JSON from a tidysummary}
\usage{
to_json(df, file = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{df}{a data frame.}

\item{file}{file name to save. If file is \code{NULL}, then the output is returned.
Otherwise, there is no output.}

\item{overwrite}{whether to overwrite the file if it already exists.}

\item{...}{other arguments passed on to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
}
\value{
A character vector of the table source code.
If file is not \code{NULL}, however, nothing is returned.
}
\description{
Takes a tidysummary object and converts it into JSON that is returned.
If the \code{file} argument is supplied, the JSON is saved to a file
instead.
}
\details{
Sends the data frame to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.
}
\examples{
\dontrun{
data <- tidysummary(iris)

to_json(data)
to_json(data, file="summary.json")
}
}
\seealso{
\code{\link[tidysummary:tidysummary]{tidysummary()}}
}
