% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_latex}
\alias{to_latex}
\title{Create a table in LaTeX from a tidysummary}
\usage{
to_latex(df, file = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{df}{a data frame.}

\item{file}{file name to save. If file is \code{NULL}, then the output is returned.
Otherwise, there is no output.}

\item{overwrite}{whether to overwrite the file if it already exists.}

\item{...}{other arguments passed on to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
A character vector of the table source code.
}
\description{
Takes a tidysummary object and converts it into LaTeX that is returned.
If the \code{file} argument is supplied, the LaTeX is saved to a file
instead.
}
\details{
Sends the data frame to \code{\link[knitr:kable]{knitr::kable()}}.
}
\examples{
\dontrun{
data <- tidysummary(iris)

to_latex(data)
to_latex(data, file="summary.latex")
}
}
\seealso{
\code{\link[tidysummary:tidysummary]{tidysummary()}}
}
