% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_markdown}
\alias{to_markdown}
\title{Create a table in Markdown from a tidysummary}
\usage{
to_markdown(df, file = NULL, overwrite = FALSE, column_quotes = FALSE, ...)
}
\arguments{
\item{df}{a data frame.}

\item{file}{file name to save. If file is \code{NULL}, then the output is returned.
Otherwise, there is no output.}

\item{overwrite}{whether to overwrite the file if it already exists.}

\item{column_quotes}{whether to use quotes around column names.}

\item{...}{other arguments passed on to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\value{
A character vector of the table source code.
}
\description{
Takes a tidysummary object and converts it into Markdown that is returned.
If the \code{file} argument is supplied, the Markdown is saved to a file
instead.
}
\details{
Sends the data frame to \code{\link[knitr:kable]{knitr::kable()}}.
}
\examples{
\dontrun{
data <- tidysummary(iris)

to_markdown(data)
to_markdown(data, file="summary.md")
}
}
\seealso{
\code{\link[tidysummary:tidysummary]{tidysummary()}}
}
